requirements.txt
discord.py
python-dotenv
asyncio
aiohttp
bs4
requests
openai
yt-dlp